+++ python-runnables/add-node-pool/runnable.py
@@ -7,7 +7,7 @@ from dku_aws.eksctl_command import EksctlCommand
 from dku_aws.aws_command import AwsCommand
 from dku_utils.cluster import get_cluster_from_dss_cluster, get_connection_info
 from dku_utils.config_parser import get_security_groups_arg, get_region_arg
-from dku_utils.node_pool import get_node_pool_args
+from dku_utils.node_pool import get_node_pool_args, build_node_pool_taints_yaml
 from dku_utils.access import _is_none_or_blank
 
 class MyRunnable(Runnable):
@@ -34,8 +34,9 @@ class MyRunnable(Runnable):
         kube_config_path = dss_cluster_settings.get_raw()['containerSettings']['executionConfigsGenericOverrides']['kubeConfigPath']
 
         connection_info = get_connection_info(dss_cluster_config.get('config'))
-        
-        node_group_id = self.config.get('nodeGroupId', None)
+
+        node_pool = self.config.get('nodePool', {})
+        node_group_id = node_pool.get('nodeGroupId', None)
         
         # first pass: get the yaml config corresponding to the command line args
         args = ['create', 'nodegroup']
@@ -54,7 +55,6 @@ class MyRunnable(Runnable):
             
         args += get_security_groups_arg(dss_cluster_config['config'].get('networkingSettings', {}))
 
-        node_pool = self.config.get('nodePool', {})
         args += get_node_pool_args(node_pool)
 
         c = EksctlCommand(args + ["--dry-run"], connection_info)
@@ -74,7 +74,16 @@ class MyRunnable(Runnable):
                 for command in commands.split('\n'):
                     if len(command.strip()) > 0:
                         node_pool_dict['preBootstrapCommands'].append(command)
-        
+
+        # Adding node pool taints on the only node pool we create which is managed:
+        node_group_taints = build_node_pool_taints_yaml(node_pool)
+        yaml_dict['managedNodeGroups'][0]['taints'] = node_group_taints
+
+        # Adding propagateASGTags to the node group if it is autoscaled.
+        # This propagates the labels/taints of the node group to the autoscaling group so that new nodes can be properly configured on creation (scaling up)
+        if node_pool.get('numNodesAutoscaling', False):
+            yaml_dict['managedNodeGroups'][0]['propagateASGTags'] = True
+
         yaml_loc = os.path.join(os.getcwd(), cluster_id +'_config.yaml')
         with open(yaml_loc, 'w') as outfile:
             yaml.dump(yaml_dict, outfile, default_flow_style=False)
@@ -90,11 +99,11 @@ class MyRunnable(Runnable):
         
         if node_pool.get('numNodesAutoscaling', False):
             logging.info("Nodegroup is autoscaling, ensuring autoscaler")
-            add_autoscaler_if_needed(cluster_id, self.config, cluster_data.get("cluster"), kube_config_path)
+            add_autoscaler_if_needed(cluster_id, self.config, cluster_data.get("cluster"), kube_config_path, node_group_taints)
             
         if node_pool.get('enableGPU', False):
             logging.info("Nodegroup is GPU-enabled, ensuring NVIDIA GPU Drivers")
-            add_gpu_driver_if_needed(self.config['clusterId'], kube_config_path, connection_info)
+            add_gpu_driver_if_needed(self.config['clusterId'], kube_config_path, connection_info, node_group_taints)
 
         args = ['get', 'nodegroup']
         #args = args + ['-v', '4']